lab3/image-classify/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
python-multipart
redis